fix: remove dead v1 code
