+++ benchmarking/run_benchmarks.py
@@ -3,23 +3,18 @@
 # SPDX-FileCopyrightText: © 2024 Tenstorrent USA, Inc.
 
 import argparse
-import json
 import logging
 import os
 import sys
-import time
-from datetime import datetime
 from pathlib import Path
 
-import jwt
-import requests
 
 # Add project root to Python path
 project_root = Path(__file__).parent.parent
 sys.path.insert(0, str(project_root))
 
 from utils.media_clients.media_client_factory import MediaClientFactory, MediaTaskType
-from utils.url_helpers import resolve_deploy_url, resolve_host_port
+from utils.url_helpers import resolve_deploy_url
 
 # Add the script's directory to the Python path
 # this for 0 setup python setup script
@@ -34,12 +29,9 @@ from benchmarking.benchmark_config import (
     powers_of_two_up_to,
     select_smoke_test_benchmark_config,
 )
-from utils.prompt_client import PromptClient
-from utils.prompt_configs import EnvironmentConfig
 from workflows.log_setup import setup_workflow_script_logger
 from workflows.model_spec import ModelSpec
 from workflows.runtime_config import RuntimeConfig
-from workflows.utils import run_command
 from workflows.workflow_config import (
     WORKFLOW_BENCHMARKS_CONFIG,
 )
@@ -50,8 +42,6 @@ from workflows.workflow_types import (
     InferenceEngine,
     ModelType,
 )
-
-from workflows.workflow_venvs import VENV_CONFIGS, STRUCTURED_OUTPUT_SCRIPT_NAME
 
 
 logger = logging.getLogger(__name__)
@@ -135,182 +125,19 @@ def parse_args():
     return ret_args
 
 
-def build_benchmark_command(
-    task,
-    benchmark_script,
-    params,
-    output_path,
-    service_port,
-    benchmark_config,
-    model_spec,
-    deploy_url: str = "http://127.0.0.1",
-):
-    run_timestamp = datetime.now().strftime("%Y-%m-%d_%H-%M-%S")
-    isl = params.isl
-    osl = params.osl
-    max_concurrency = params.max_concurrency
-    num_prompts = params.num_prompts
-
-    # VLM models: use isl/osl + image dimensions
-    if params.task_type == "vlm":
-        result_filename = (
-            Path(output_path)
-            / f"benchmark_{model_spec.model_id}_{run_timestamp}_isl-{isl}_osl-{osl}_maxcon-{max_concurrency}_n-{num_prompts}_images-{params.images_per_prompt}_height-{params.image_height}_width-{params.image_width}.json"
-        )
-    # Text models: standard isl/osl
-    else:
-        result_filename = (
-            Path(output_path)
-            / f"benchmark_{model_spec.model_id}_{run_timestamp}_isl-{isl}_osl-{osl}_maxcon-{max_concurrency}_n-{num_prompts}.json"
-        )
-
-    # VLM models need multimodal dataset; text models use standard dataset
-    dataset_name = "random-mm" if params.task_type == "vlm" else "random"
-    backend = "openai-chat"
-
-    host, port = resolve_host_port(deploy_url, service_port)
-
-    # fmt: off
-    cmd = [
-        str(benchmark_script),
-        "bench",
-        "serve",
-        "--backend", backend,
-        "--endpoint", "/v1/chat/completions",
-        "--model", model_spec.hf_model_repo,
-        "--host", host,
-        "--port", port,
-        "--dataset-name", dataset_name,
-        "--max-concurrency", str(max_concurrency),
-        "--num-prompts", str(num_prompts),
-        "--random-input-len", str(isl),
-        "--random-output-len", str(osl),
-        "--percentile-metrics", "ttft,tpot,itl,e2el",  # must add e2el in order for it to be logged
-        "--save-result",
-        "--save-detailed",
-        "--result-filename", str(result_filename),
-    ]
-
-    # only truncate prompts for text-only tasks; VLMs interleave vision tokens
-    # in the prompt and truncation can drop them, causing 400s at the preprocessor
-    if params.task_type == "text":
-        # Pin output length via max_tokens in the request body. --random-output-len
-        # isn't always enforced as a hard cap server-side (the forge OpenAI-compat
-        # server ignores it), so requests can run unbounded -- Qwen3-32B then emits
-        # ~1000+ reasoning tokens for osl=128 and blows the 6h CI cap. max_tokens in
-        # the body is honored.
-        cmd.extend([
-            "--extra-body", json.dumps(
-                {"truncate_prompt_tokens": str(isl), "max_tokens": int(osl)}
-            ),
-        ])
-
-    if params.task_type == "vlm":
-        if params.image_height and params.image_width:
-            cmd.extend([
-                "--random-mm-base-items-per-request", str(params.images_per_prompt),
-                "--random-mm-bucket-config", str({(params.image_height, params.image_width, 1): 1.0}),
-                "--endpoint", "/v1/chat/completions"
-            ])
-    # fmt: on
-    return cmd
-
-
-def build_structured_output_command(
-    venv_python,
-    structured_output_script,
-    params,
-    output_path,
-    service_port,
-    model_spec,
-    deploy_url: str = "http://127.0.0.1",
-):
-    run_timestamp = datetime.now().strftime("%Y-%m-%d_%H-%M-%S")
-    ratio_tag = (
-        "no-so"
-        if params.structured_output_ratio is None
-        else f"so-{params.structured_output_ratio}"
-    )
-    result_filepath = (
-        Path(output_path)
-        / f"benchmark_structured_{model_spec.model_id}_{run_timestamp}"
-        f"_dataset-{params.structured_dataset}_{ratio_tag}"
-        f"_osl-{params.osl}_maxcon-{params.max_concurrency}_n-{params.num_prompts}.json"
-    )
-
-    host, port = resolve_host_port(deploy_url, service_port)
-
-    # fmt: off
-    cmd = [
-        str(venv_python),
-        str(structured_output_script),
-        "--backend", "vllm",
-        "--host", host,
-        "--port", port,
-        "--model", model_spec.hf_model_repo,
-        "--dataset", params.structured_dataset,
-        "--max-concurrency", str(params.max_concurrency),
-        "--num-prompts", str(params.num_prompts),
-        "--output-len", str(params.osl),
-        "--request-rate", "inf",
-        "--percentile-metrics", "ttft,tpot,itl,e2el",
-        "--save-results",
-        "--result-dir", str(output_path),
-        "--result-filename", result_filepath.name,
-    ]
-    if params.structured_output_ratio is None:
-        cmd.append("--no-structured-output")
-    else:
-        cmd.extend(["--structured-output-ratio", str(params.structured_output_ratio)])
-    # fmt: on
-    return cmd, result_filepath
-
-
-def annotate_structured_output_result(result_filepath):
-    """Tag a structured-output result JSON in place so downstream readers can
-    distinguish it from regular `vllm bench serve` outputs."""
-    if not result_filepath.exists():
-        logger.warning("Structured-output result JSON not found: %s", result_filepath)
-        return
-    try:
-        with open(result_filepath, "r+") as f:
-            data = json.load(f)
-            data["benchmark_kind"] = "structured_output"
-            f.seek(0)
-            f.truncate()
-            json.dump(data, f, indent=2)
-    except (OSError, json.JSONDecodeError) as exc:
-        logger.warning("Could not annotate %s: %s", result_filepath, exc)
-
-
 def main():
     # Setup logging configuration.
     setup_workflow_script_logger(logger)
     logger.info(f"Running {__file__} ...")
 
     args = parse_args()
-    jwt_secret = args.jwt_secret
     model_spec = ModelSpec.from_json(args.runtime_model_spec_json)
     runtime_config = RuntimeConfig.from_json(args.runtime_model_spec_json)
 
     # runtime config loaded from JSON
     device_str = runtime_config.device
     service_port = runtime_config.service_port
-    disable_trace_capture = runtime_config.disable_trace_capture
-    # Resolved once here so the genai-perf branch (which returns before
-    # env_config is constructed) sees the same value used later.
     deploy_url = resolve_deploy_url(runtime_config)
-
-    # Automatically control trace capture based on has_builtin_warmup
-    # Only apply automatic logic if user hasn't explicitly set --disable-trace-capture
-    if not disable_trace_capture and hasattr(model_spec, "has_builtin_warmup"):
-        if model_spec.has_builtin_warmup:
-            # Model has builtin warmup - disable client-side trace capture
-            disable_trace_capture = True
-            logger.info(
-                "Model has builtin warmup (has_builtin_warmup=True), "
-                "automatically disabling trace capture for benchmarks workflow"
-            )
 
     device = DeviceTypes.from_string(device_str)
     workflow_config = WORKFLOW_BENCHMARKS_CONFIG
@@ -321,16 +148,6 @@ def main():
     logger.info(f"output_path=: {args.output_path}")
 
     # set environment vars
-    if jwt_secret:
-        # If jwt-secret is provided, generate the JWT and set OPENAI_API_KEY.
-        json_payload = json.loads(
-            '{"team_id": "tenstorrent", "token_id": "debug-test"}'
-        )
-        encoded_jwt = jwt.encode(json_payload, jwt_secret, algorithm="HS256")
-        os.environ["OPENAI_API_KEY"] = encoded_jwt
-        logger.info(
-            "OPENAI_API_KEY environment variable set using provided JWT secret."
-        )
     if (
         model_spec.inference_engine == InferenceEngine.MEDIA.value
         or model_spec.inference_engine == InferenceEngine.FORGE.value
@@ -338,8 +155,6 @@ def main():
         os.environ["OPENAI_API_KEY"] = "your-secret-key"
         os.environ["VLLM_API_KEY"] = "your-secret-key"
         logger.info("VLLM_API_KEY environment variable set to your-secret-key.")
-
-    env_vars = os.environ.copy()
 
     # Look up the evaluation configuration for the model using BENCHMARK_CONFIGS.
     if model_spec.model_id not in BENCHMARK_CONFIGS:
@@ -406,151 +221,15 @@ def main():
         )
         return return_code
 
-    log_str = "Running benchmarks for:\n"
-    log_str += f"  {'#':<3} {'isl':<10} {'osl':<10} {'max_concurrency':<15} {'num_prompts':<12}\n"
-    log_str += (
-        f"  {'-' * 3:<3} {'-' * 10:<10} {'-' * 10:<10} {'-' * 15:<15} {'-' * 12:<12}\n"
+    # LLM benchmarks (incl. prefix-cache / spec-decode / serving-bench) are
+    # served by the v2 engine (tt-inference-server-v2 via workflows/v2_bridge);
+    # the v1 vLLM/genai-perf benchmark path has been retired. Reaching here means
+    # can_route_to_v2 did not route this model to v2 (a routing regression).
+    raise SystemExit(
+        f"benchmarks for {model_spec.model_name!r} are served by the v2 engine; "
+        "the v1 benchmark path has been retired. Reaching this branch means "
+        "workflows/v2_bridge.can_route_to_v2 did not route this model to v2."
     )
-    for i, param in enumerate(all_params, 1):
-        if param.task_type == "text":
-            log_str += f"  {i:<3} {param.isl:<10} {param.osl:<10} {param.max_concurrency:<15} {param.num_prompts:<12}\n"
-    if "image" in model_spec.supported_modalities:
-        log_str += "Running VLM benchmarks for:\n"
-        log_str += f"  {'#':<3} {'isl':<10} {'osl':<10} {'max_concurrency':<15} {'images_per_prompt':<12} {'image_height':<12} {'image_width':<12} {'num_prompts':<12}\n"
-        log_str += f"  {'-' * 3:<3} {'-' * 10:<10} {'-' * 10:<10} {'-' * 15:<15} {'-' * 12:<12} {'-' * 12:<12} {'-' * 12:<12} {'-' * 12:<12}\n"
-        for i, param in enumerate(all_params, 1):
-            if param.task_type == "vlm":
-                log_str += f"  {i:<3} {param.isl:<10} {param.osl:<10} {param.max_concurrency:<15} {param.images_per_prompt:<12} {param.image_height:<12} {param.image_width:<12} {param.num_prompts:<12}\n"
-    logger.info(log_str)
-
-    if not all_params:
-        message = f"No benchmark tasks defined for model: {model_spec.model_name} on device: {device.name}"
-        raise AssertionError(message)
-
-    logger.info("Wait for the vLLM server to be ready ...")
-    env_config = EnvironmentConfig()
-    env_config.jwt_secret = jwt_secret
-    env_config.vllm_api_key = os.getenv("VLLM_API_KEY")
-    env_config.service_port = service_port
-    env_config.vllm_model = model_spec.hf_model_repo
-    env_config.deploy_url = deploy_url
-
-    prompt_client = PromptClient(
-        env_config,
-        model_spec=model_spec,
-        runtime_config=runtime_config,
-    )
-    logger.info(
-        "Using tensor_cache_timeout:=%ss as the vLLM server startup budget (covers tensor cache generation and warm-cache restarts on multi-DP-engine deployments)",
-        prompt_client.cache_monitor.get_tensor_cache_timeout(),
-    )
-    if not prompt_client.wait_for_healthy():
-        logger.error("⛔️ vLLM server is not healthy. Aborting benchmarks. ")
-        return 1
-
-    # keep track of captured traces to avoid re-running requests
-    captured_traces = set()
-
-    # Run benchmarks
-    return_codes = []
-    for task in benchmark_config.tasks:
-        venv_config = VENV_CONFIGS[task.workflow_venv_type]
-        is_structured_output_task = (
-            task.task_type == BenchmarkTaskType.HTTP_CLIENT_VLLM_STRUCTURED_OUTPUT_API
-        )
-        benchmark_script = venv_config.venv_path / "bin" / "vllm"
-        if device in task.param_map:
-            params_list = task.param_map[device]
-            # Structured-output params have no isl, so they don't participate in
-            # trace capture. The structured-output script generates its own
-            # prompt shapes from the chosen dataset.
-            context_lens_set = {
-                (params.isl, params.osl)
-                for params in params_list
-                if params.isl is not None and params.osl is not None
-            }
-            context_lens_set.difference_update(captured_traces)
-            # ascending order of input sequence length
-            sorted_context_lens_set = sorted(context_lens_set)
-            if not disable_trace_capture and sorted_context_lens_set:
-                if "image" in model_spec.supported_modalities:
-                    prompt_client.capture_traces(
-                        context_lens=list(sorted_context_lens_set),
-                        timeout=1200.0,
-                        image_resolutions=IMAGE_RESOLUTIONS,
-                    )
-                else:
-                    prompt_client.capture_traces(
-                        context_lens=list(sorted_context_lens_set), timeout=1200.0
-                    )
-                captured_traces.update(sorted_context_lens_set)
-            for i, params in enumerate(params_list, 1):
-                try:
-                    health_check = prompt_client.get_health()
-                except requests.exceptions.RequestException as error:
-                    logger.error("Health check request failed: %s", error)
-                    return 1
-                if health_check.status_code != 200:
-                    logger.error("⛔️ vLLM server is not healthy. Aborting benchmarks.")
-                    return 1
-
-                logger.info(
-                    f"Running benchmark {model_spec.model_name}: {i}/{len(params_list)}"
-                )
-                # Add a small delay between runs to ensure system stability
-                time.sleep(2)
-                structured_output_result_filepath = None
-                if is_structured_output_task:
-                    structured_output_script = (
-                        venv_config.venv_path
-                        / "work_dir"
-                        / STRUCTURED_OUTPUT_SCRIPT_NAME
-                    )
-                    if not structured_output_script.exists():
-                        logger.error(
-                            "⛔️ Structured-output benchmark script not found at "
-                            f"{structured_output_script}. Re-run with --reset-venvs "
-                            "or rerun fetch_structured_output_scripts to fetch it."
-                        )
-                        return_codes.append(1)
-                        continue
-                    cmd, structured_output_result_filepath = (
-                        build_structured_output_command(
-                            venv_python=venv_config.venv_python,
-                            structured_output_script=structured_output_script,
-                            params=params,
-                            output_path=args.output_path,
-                            service_port=service_port,
-                            model_spec=model_spec,
-                            deploy_url=deploy_url,
-                        )
-                    )
-                else:
-                    cmd = build_benchmark_command(
-                        task,
-                        benchmark_script,
-                        params=params,
-                        output_path=args.output_path,
-                        service_port=service_port,
-                        benchmark_config=benchmark_config,
-                        model_spec=model_spec,
-                        deploy_url=deploy_url,
-                    )
-                return_code = run_command(command=cmd, logger=logger, env=env_vars)
-                return_codes.append(return_code)
-                if return_code == 0 and structured_output_result_filepath is not None:
-                    annotate_structured_output_result(structured_output_result_filepath)
-
-    if all(return_code == 0 for return_code in return_codes):
-        logger.info("✅ Completed benchmarks")
-        main_return_code = 0
-    else:
-        logger.error(
-            f"⛔ benchmarks failed with return codes: {return_codes}. See logs above for details."
-        )
-        main_return_code = 1
-
-    return main_return_code
 
 
 def run_benchmarks(
